This commit is fixing a commit that wiped out basically all my changes from my last few commits.
A hover effect was added to MainMenuCSS.css. Multiple images were deleted that weren't being used anymore.

Leaderboard methods updated in MainMode ready to increment the score for a given player, as well as commenting out the reset method inside of the game loop.

Score changed in ClientController from 3 to 1 during testing phases.

Changes undone to SceneManager from previous commit.

Undid changes made to ApplicationUser.username being made static as there is already a static reference to the object created in MenuController which can be accessed if needs be.

Created MenuUtilitites to deal with code which was being repeated throughout the codebase. Currently contains a method to get a background given a class reference and a string of the background location.

Added Maps reference inside of AI to allow the AI to be aware of the map.

Added explanation comments to Node and removed unnecessary white space from the end of the file.

Reverted changes made to Controllable from the last commit.

Changed all references to javax and awt to JavaFX instead (as we really shouldn't be mixing these frameworks) inside of Server and reformatted the whole file.

MapRender changed to use map reference stored in Maps.map.

Deleted lots of unnecessary files: Database.java, game_Menu.fxml, GameContainer.java, GameMenu.java, GameWindow.java, half-life.fxml, HalfLifeView.java, Window.java